✨ Renamed ID Generation
